Rename app validate API module to build_validate_api
The build-trigger endpoint being added next lives alongside the
validate endpoints, so validate.py no longer describes the file.
